Addition of distributed app, and moving of additionApp (single console) to its own folder too.
This now creates two different solutions in the repository. AddingApp has not changed from the previous submission. DistributedApp is an modification of AddingApp, where two consoles are loaded, one to receive input through console and submit to MSMQ, another console to receive off MSMQ and print the calculated sum of the objects on the queue. The application also contains SpecFlow tests, demonstrating the capability.